Refactor common legal components
