+++ app/src/components/LegalComponents/LegalComponents.css
@@ -1,0 +1,7 @@
+.list-clause-header {
+  list-style: none;
+}
+
+.bold {
+  font-weight: bold;
+}
